requirements.txt
agno==2.3.12
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
brotli==1.2.0
certifi==2025.11.12
click==8.3.1
colorama==0.4.6
ddgs==9.9.3
distro==1.9.0
docstring_parser==0.17.0
duckduckgo_search==8.1.1
fake-useragent==2.2.0
fastapi==0.124.4
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.3.0
groq==0.37.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
jiter==0.12.0
lxml==6.0.2
markdown-it-py==4.0.0
mdurl==0.1.2
openai==2.11.0
packaging==25.0
primp==0.15.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
rich==14.2.0
shellingham==1.5.4
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
SQLAlchemy==2.0.45
sqlmodel==0.0.27
starlette==0.50.0
tqdm==4.67.1
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
